refactor(agent): split harness/archivist.rs into archivist/ directory
Splits the 1257-line archivist.rs into focused sub-modules each ≤500 lines:
types.rs (struct def), lifecycle.rs (constructors + segment ops + on_segment_closed),
recap.rs (summarize_entries + rolling_segment_recap + read_session_entries),
tree_ingest.rs (pipe_segment_to_tree), hook_impl.rs (PostTurnHook impl),
helpers.rs (utility fns), and test_constructors.rs (#[cfg(test)] impls).
External API preserved via pub use re-export in mod.rs.
