+++ src/openhuman/config/schema/tools/browser.rs
@@ -1,0 +1,96 @@
+//! Browser and computer-use config types.
+
+use super::super::defaults;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[serde(default)]
+pub struct BrowserComputerUseConfig {
+    #[serde(default = "default_browser_computer_use_endpoint")]
+    pub endpoint: String,
+    #[serde(default = "default_browser_computer_use_timeout_ms")]
+    pub timeout_ms: u64,
+    #[serde(default)]
+    pub allow_remote_endpoint: bool,
+    #[serde(default)]
+    pub window_allowlist: Vec<String>,
+    #[serde(default)]
+    pub max_coordinate_x: Option<i64>,
+    #[serde(default)]
+    pub max_coordinate_y: Option<i64>,
+}
+
+fn default_browser_computer_use_endpoint() -> String {
+    "http://127.0.0.1:8787/v1/actions".into()
+}
+
+fn default_browser_computer_use_timeout_ms() -> u64 {
+    15_000
+}
+
+impl Default for BrowserComputerUseConfig {
+    fn default() -> Self {
+        Self {
+            endpoint: default_browser_computer_use_endpoint(),
+            timeout_ms: default_browser_computer_use_timeout_ms(),
+            allow_remote_endpoint: false,
+            window_allowlist: Vec::new(),
+            max_coordinate_x: None,
+            max_coordinate_y: None,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[serde(default)]
+pub struct BrowserConfig {
+    #[serde(default)]
+    pub enabled: bool,
+    /// DEPRECATED: the browser tool now shares the unified web-access host list
+    /// in `[http_request].allowed_domains` (see `tools::ops::all_tools_with_runtime`).
+    /// Still parsed for backward compatibility but no longer gates browser
+    /// navigation. Manage allowed hosts via Settings → Search → Allowed websites;
+    /// browser allow-all remains gated by `OPENHUMAN_BROWSER_ALLOW_ALL`.
+    #[serde(default)]
+    pub allowed_domains: Vec<String>,
+    #[serde(default)]
+    pub session_name: Option<String>,
+    #[serde(default = "default_browser_backend")]
+    pub backend: String,
+    #[serde(default = "default_true")]
+    pub native_headless: bool,
+    #[serde(default = "default_browser_webdriver_url")]
+    pub native_webdriver_url: String,
+    #[serde(default)]
+    pub native_chrome_path: Option<String>,
+    #[serde(default)]
+    pub computer_use: BrowserComputerUseConfig,
+}
+
+fn default_true() -> bool {
+    defaults::default_true()
+}
+
+fn default_browser_backend() -> String {
+    "agent_browser".into()
+}
+
+fn default_browser_webdriver_url() -> String {
+    "http://127.0.0.1:9515".into()
+}
+
+impl Default for BrowserConfig {
+    fn default() -> Self {
+        Self {
+            enabled: false,
+            allowed_domains: Vec::new(),
+            session_name: None,
+            backend: default_browser_backend(),
+            native_headless: default_true(),
+            native_webdriver_url: default_browser_webdriver_url(),
+            native_chrome_path: None,
+            computer_use: BrowserComputerUseConfig::default(),
+        }
+    }
+}
